test(onboard): split policy helper coverage
